Integrator thread now a child of integratorTree
All code related to integration now contained within integrator
Clock is once again implemented to ensure only appropriate
parts of the program are enabled for the user depending
on the state.
